+++ gcloud-manager/src/main/java/core/infra/command/SyncDBCommand.java
@@ -6,6 +6,7 @@ import core.infra.gcloud.SQLInstanceClient;
 import core.infra.gcloud.SecretClient;
 import core.infra.kube.KubeClient;
 import core.infra.util.JSON;
+import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
@@ -20,20 +21,22 @@ import java.util.Set;
  * @author neo
  */
 public class SyncDBCommand {
+    private final Logger logger = LoggerFactory.getLogger(SyncDBCommand.class);
     private final SQLInstanceClient sqlInstance = new SQLInstanceClient();
     private final SecretClient secretClient = new SecretClient();
     private final KubeClient kubeClient = new KubeClient();
     private final DBConfig config;
 
     public SyncDBCommand(Path configPath) throws IOException {
-        LoggerFactory.getLogger(SyncDBCommand.class).info("sync db, config={}", configPath.toAbsolutePath());
+        logger.info("load db config, config={}", configPath.toAbsolutePath());
         config = JSON.fromJSON(DBConfig.class, Files.readString(configPath));
         config.validate();
     }
 
     public void sync() throws Exception {
+        logger.info("sync db");
         DescribeSQLResponse instance = sqlInstance.describe(config.project, config.instance);
-        String rootPassword = secretClient.getOrCreateSecret(config.project, secretName(config.rootSecret), config.env);
+        String rootPassword = secretClient.getOrCreateSecret(config.project, config.rootSecret, config.env);
         sqlInstance.changeRootPassword(config.project, config.instance, rootPassword);
         kubeClient.switchContext(config.project, config.kube.name, config.kube.zone);
 
@@ -43,7 +46,7 @@ public class SyncDBCommand {
                 client.createDB(db);
             }
             for (DBConfig.User user : config.users) {
-                String password = secretClient.getOrCreateSecret(config.project, secretName(user.secret), config.env);
+                String password = secretClient.getOrCreateSecret(config.project, user.secret, config.env);
                 createDBUser(client, user, password);
                 if (user.kube != null) {
                     if (namespaces.add(user.kube.ns)) {
@@ -68,9 +71,5 @@ public class SyncDBCommand {
             case "READ_ONLY" -> client.createUser(user.name, password, "*", List.of("SELECT"));
             default -> throw new Error("unknown user type, type=" + user.type);
         }
-    }
-
-    private String secretName(String name) {
-        return config.env + "-" + name;
     }
 }