+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ### 0.0.2
 
-* add "env" on db config, to label secret and as prefix of secret
+* add "env" on db config, to label secret
 * add "db-migrate-secret-label" command, to add labels to existing secret